+++ Package.swift
@@ -1,29 +1,17 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.3
 import PackageDescription
 
 let package = Package(
     name: "Microya",
+    platforms: [.macOS(.v10_15), .iOS(.v13), .tvOS(.v13)],
     products: [
         .library(name: "Microya", targets: ["Microya"])
     ],
-    dependencies: [
-//        .package(url: "https://github.com/Flinesoft/HandySwift.git", .upToNextMajor(from: "2.5.0")),
-//        .package(url: "https://github.com/Flinesoft/HandyUIKit.git", .upToNextMajor(from: "1.6.0"))
-    ],
     targets: [
-        .target(
-            name: "Microya",
-            dependencies: [
-//                "HandySwift",
-//                "HandyUIKit"
-            ],
-            path: "Frameworks/Microya",
-            exclude: ["Frameworks/SupportingFiles"]
-        ),
+        .target(name: "Microya"),
         .testTarget(
             name: "MicroyaTests",
-            dependencies: ["Microya"],
-            exclude: ["Tests/SupportingFiles"]
+            dependencies: ["Microya"]
         )
     ]
 )